pacman/requirements-dev.txt
chalice
pytest

mysql-connector-python
boto3
